Make nskip an integer

+++ pylal/pylal/bayespputils.py
@@ -5555,13 +5555,12 @@ def find_ndownsample(samples, nDownsample):
     nskip = 1
     if nDownsample is not None:
         if len(samples) > nDownsample:
-            nskip *= floor(len(samples)/nDownsample)
-
+            nskip *= int(floor(len(samples)/nDownsample))
     else:
         nEff = nEffective
         if nEff > 1:
             if len(samples) > nEff:
-                nskip = ceil(len(samples)/nEff)
+                nskip = int(ceil(len(samples)/nEff))
         else:
             nskip = np.nan
     return nskip
@@ -5821,15 +5820,14 @@ class PEOutputParser(object):
         ntot = sum(ntots)
         if nDownsample is not None:
             if ntot > nDownsample:
-                nskips *= floor(ntot/nDownsample)
-
+                nskips *= int(floor(ntot/nDownsample))
         else:
             for i in range(nfiles):
                 nEff = nEffectives[i]
                 ntot = ntots[i]
                 if nEff > 1:
                     if ntot > nEff:
-                        nskips[i] = ceil(ntot/nEff)
+                        nskips[i] = int(ceil(ntot/nEff))
                 else:
                     nskips[i] = None
         return nskips
